LLVM.Internal.Context

Documentation

data Context

a Context object holds the state the of LLVM system needs for one thread of | LLVM compilation. Once upon a time, in early versions of LLVM, this state was global. | Then it got packed up in this object to allow multiple threads to compile at once.

Constructors

Context (Ptr Context)

withContext :: (Context -> IO a) -> IO a

Create a Context, run an action (to which it is provided), then destroy the Context.