LLVM.Internal.FFI.Assembly

Description

Functions to read and write textual LLVM assembly

Documentation

parseLLVMAssembly :: Ptr Context -> OwnerTransfered (Ptr MemoryBuffer) -> Ptr (OwnerTransfered CString) -> IO (Ptr Module)

Use LLVM's parser to parse a string of llvm assembly in a memory buffer to get a module

writeLLVMAssembly :: Ptr Module -> Ptr RawOStream -> IO ()

LLVM's serializer to generate a string of llvm assembly from a module