LLVM.Internal.FFI.Cleanup

Documentation

foreignDecl :: String -> String -> [TypeQ] -> TypeQ -> DecsQ

hasFlags :: [Type] -> Bool

The LLVM C-API for instructions with boolean flags (e.g. nsw) and is weak, so they get separated out for different handling. This check is an accurate but crude test for whether an instruction needs such handling.

typeMapping :: Type -> TypeQ