LLVM.Internal.FFI.Iterate

Description

Functions to help handle LLVM iteration patterns

Documentation

getXs :: IO (Ptr a) -> (Ptr a -> IO (Ptr a)) -> IO [Ptr a]

retrieve a sequence of objects which form a linked list, given an action to | retrieve the first member and an action to proceed through the list