LLVM.Internal.FFI.LLVMCTypes

Description

Define types which correspond cleanly with some simple types on the C/C++ side. Encapsulate hsc macro weirdness here, supporting higher-level tricks elsewhere.

Documentation

newtype LLVMBool

Constructors

LLVMBool CUInt

newtype OwnerTransfered a

If an FFI function returns a value wrapped in OwnerTransfered, this value needs to be freed after it has been processed. Usually this is done automatically in the DecodeM instance.

Constructors

OwnerTransfered a

newtype NothingAsMinusOne h

Constructors

NothingAsMinusOne CInt

newtype NothingAsEmptyString c

Constructors

NothingAsEmptyString c

newtype CPPOpcode

Constructors

CPPOpcode CUInt

newtype ICmpPredicate

Constructors

ICmpPredicate CUInt

iCmpPredEQ :: ICmpPredicate

iCmpPredNE :: ICmpPredicate

iCmpPredUGT :: ICmpPredicate

iCmpPredUGE :: ICmpPredicate

iCmpPredULT :: ICmpPredicate

iCmpPredULE :: ICmpPredicate

iCmpPredSGT :: ICmpPredicate

iCmpPredSGE :: ICmpPredicate

iCmpPredSLT :: ICmpPredicate

iCmpPredSLE :: ICmpPredicate

newtype FCmpPredicate

Constructors

FCmpPredicate CUInt

fCmpPredFalse :: FCmpPredicate

fCmpPredOEQ :: FCmpPredicate

fCmpPredOGT :: FCmpPredicate

fCmpPredOGE :: FCmpPredicate

fCmpPredOLT :: FCmpPredicate

fCmpPredOLE :: FCmpPredicate

fCmpPredONE :: FCmpPredicate

fCmpPredORD :: FCmpPredicate

fCmpPredUNO :: FCmpPredicate

fCmpPredUEQ :: FCmpPredicate

newtype MDKindID

Constructors

MDKindID CUInt

fCmpPredUGT :: FCmpPredicate

fCmpPredUGE :: FCmpPredicate

newtype FastMathFlags

Constructors

FastMathFlags CUInt

fCmpPredULT :: FCmpPredicate

fastMathFlagsUnsafeAlgebra :: FastMathFlags

fCmpPredULE :: FCmpPredicate

fastMathFlagsNoNaNs :: FastMathFlags

newtype MemoryOrdering

Constructors

MemoryOrdering CUInt

fCmpPredUNE :: FCmpPredicate

fastMathFlagsNoInfs :: FastMathFlags

fcmpPredTrue :: FCmpPredicate

memoryOrderingNotAtomic :: MemoryOrdering

fastMathFlagsNoSignedZeros :: FastMathFlags

memoryOrderingUnordered :: MemoryOrdering

newtype UnnamedAddr

Constructors

UnnamedAddr CUInt

fastMathFlagsAllowReciprocal :: FastMathFlags

memoryOrderingMonotonic :: MemoryOrdering

unnamedAddrNone :: UnnamedAddr

fastMathFlagsAllowContract :: FastMathFlags

memoryOrderingAcquire :: MemoryOrdering

fastMathFlagsP :: QuasiQuoter

unnamedAddrLocal :: UnnamedAddr

newtype SynchronizationScope

Constructors

SynchronizationScope CUInt

memoryOrderingRelease :: MemoryOrdering

unnamedAddrGlobal :: UnnamedAddr

memoryOrderingAcquireRelease :: MemoryOrdering

synchronizationScopeSingleThread :: SynchronizationScope

unnamedAddrP :: QuasiQuoter

synchronizationScopeSystem :: SynchronizationScope

memoryOrderingSequentiallyConsistent :: MemoryOrdering

newtype TailCallKind

Constructors

TailCallKind CUInt

memoryOrderingP :: QuasiQuoter

synchronizationScopeP :: QuasiQuoter

tailCallKindNone :: TailCallKind

tailCallKindTail :: TailCallKind

newtype Linkage

Constructors

Linkage CUInt

tailCallKindMustTail :: TailCallKind

linkageExternal :: Linkage

tailCallKindP :: QuasiQuoter

linkageAvailableExternally :: Linkage

newtype Visibility

Constructors

Visibility CUInt

linkageLinkOnceAny :: Linkage

visibilityDefault :: Visibility

linkageLinkOnceODR :: Linkage

visibilityHidden :: Visibility

newtype COMDATSelectionKind

Constructors

COMDATSelectionKind CUInt

linkageWeakAny :: Linkage

visibilityProtected :: Visibility

linkageWeakODR :: Linkage

comdatSelectionKindAny :: COMDATSelectionKind

visibilityP :: QuasiQuoter

linkageAppending :: Linkage

comdatSelectionKindExactMatch :: COMDATSelectionKind

newtype DLLStorageClass

Constructors

DLLStorageClass CUInt

linkageInternal :: Linkage

comdatSelectionKindLargest :: COMDATSelectionKind

dllStorageClassDefault :: DLLStorageClass

linkagePrivate :: Linkage

comdatSelectionKindNoDuplicates :: COMDATSelectionKind

dllStorageClassDLLImport :: DLLStorageClass

newtype CallingConvention

Constructors

CallingConvention CUInt

linkageExternalWeak :: Linkage

comdatSelectionKindSameSize :: COMDATSelectionKind

dllStorageClassDLLExport :: DLLStorageClass

linkageCommon :: Linkage

comdatSelectionKindP :: QuasiQuoter

callingConventionC :: CallingConvention

dllStorageClassP :: QuasiQuoter

linkageP :: QuasiQuoter

callingConventionFast :: CallingConvention

newtype ThreadLocalMode

Constructors

ThreadLocalMode CUInt

callingConventionCold :: CallingConvention

threadLocalModeNotThreadLocal :: ThreadLocalMode

callingConventionGHC :: CallingConvention

threadLocalModeGeneralDynamicTLSModel :: ThreadLocalMode

newtype ValueSubclassId

Constructors

ValueSubclassId CUInt

callingConventionHiPE :: CallingConvention

threadLocalModeLocalDynamicTLSModel :: ThreadLocalMode

valueSubclassIdArgument :: ValueSubclassId

callingConventionWebKit_JS :: CallingConvention

threadLocalModeInitialExecTLSModel :: ThreadLocalMode

valueSubclassIdBasicBlock :: ValueSubclassId

callingConventionAnyReg :: CallingConvention

newtype DiagnosticKind

Constructors

DiagnosticKind CUInt

threadLocalModeLocalExecTLSModel :: ThreadLocalMode

valueSubclassIdFunction :: ValueSubclassId

callingConventionPreserveMost :: CallingConvention

threadLocalModeP :: QuasiQuoter

diagnosticKindError :: DiagnosticKind

valueSubclassIdGlobalAlias :: ValueSubclassId

callingConventionPreserveAll :: CallingConvention

diagnosticKindWarning :: DiagnosticKind

newtype AsmDialect

Constructors

AsmDialect CUInt

valueSubclassIdGlobalVariable :: ValueSubclassId

callingConventionX86_StdCall :: CallingConvention

diagnosticKindNote :: DiagnosticKind

asmDialectATT :: AsmDialect

valueSubclassIdUndefValue :: ValueSubclassId

callingConventionX86_FastCall :: CallingConvention

diagnosticKindP :: QuasiQuoter

asmDialectIntel :: AsmDialect

valueSubclassIdBlockAddress :: ValueSubclassId

callingConventionARM_APCS :: CallingConvention

newtype RMWOperation

Constructors

RMWOperation CUInt

asmDialectP :: QuasiQuoter

valueSubclassIdConstantExpr :: ValueSubclassId

callingConventionARM_AAPCS :: CallingConvention

rmwOperationXchg :: RMWOperation

callingConventionARM_AAPCS_VFP :: CallingConvention

valueSubclassIdConstantAggregateZero :: ValueSubclassId

rmwOperationAdd :: RMWOperation

newtype RelocModel

Constructors

RelocModel CUInt

callingConventionMSP430_INTR :: CallingConvention

valueSubclassIdConstantDataArray :: ValueSubclassId

rmwOperationSub :: RMWOperation

relocModelDefault :: RelocModel

callingConventionX86_ThisCall :: CallingConvention

valueSubclassIdConstantDataVector :: ValueSubclassId

rmwOperationAnd :: RMWOperation

relocModelStatic :: RelocModel

valueSubclassIdConstantInt :: ValueSubclassId

callingConventionPTX_Kernel :: CallingConvention

newtype CodeModel

Constructors

CodeModel CUInt

rmwOperationNand :: RMWOperation

relocModelPIC :: RelocModel

valueSubclassIdConstantFP :: ValueSubclassId

callingConventionPTX_Device :: CallingConvention

codeModelDefault :: CodeModel

rmwOperationOr :: RMWOperation

relocModelDynamicNoPic :: RelocModel

valueSubclassIdConstantArray :: ValueSubclassId

callingConventionSPIR_FUNC :: CallingConvention

rmwOperationXor :: RMWOperation

codeModelJITDefault :: CodeModel

newtype CodeGenOptLevel

Constructors

CodeGenOptLevel CUInt

relocModelP :: QuasiQuoter

valueSubclassIdConstantStruct :: ValueSubclassId

callingConventionSPIR_KERNEL :: CallingConvention

codeModelSmall :: CodeModel

rmwOperationMax :: RMWOperation

codeGenOptLevelNone :: CodeGenOptLevel

callingConventionIntel_OCL_BI :: CallingConvention

valueSubclassIdConstantTokenNone :: ValueSubclassId

codeModelKernel :: CodeModel

rmwOperationMin :: RMWOperation

codeGenOptLevelLess :: CodeGenOptLevel

valueSubclassIdConstantVector :: ValueSubclassId

callingConventionX86_64_SysV :: CallingConvention

newtype CodeGenFileType

Constructors

CodeGenFileType CUInt

codeModelMedium :: CodeModel

rmwOperationUMax :: RMWOperation

codeGenOptLevelDefault :: CodeGenOptLevel

callingConventionWin64 :: CallingConvention

valueSubclassIdConstantPointerNull :: ValueSubclassId

codeModelLarge :: CodeModel

rmwOperationUMin :: RMWOperation

codeGenFileTypeAssembly :: CodeGenFileType

callingConventionP :: QuasiQuoter

valueSubclassIdInlineAsm :: ValueSubclassId

codeGenOptLevelAggressive :: CodeGenOptLevel

codeModelP :: QuasiQuoter

rmwOperationP :: QuasiQuoter

codeGenFileTypeObject :: CodeGenFileType

newtype FloatABIType

Constructors

FloatABIType CUInt

codeGenOptLevelP :: QuasiQuoter

valueSubclassIdInstruction :: ValueSubclassId

codeGenFileTypeP :: QuasiQuoter

floatABIDefault :: FloatABIType

valueSubclassIdP :: QuasiQuoter

floatABISoft :: FloatABIType

newtype FPOpFusionMode

Constructors

FPOpFusionMode CUInt

floatABIHard :: FloatABIType

fpOpFusionModeFast :: FPOpFusionMode

floatABIP :: QuasiQuoter

fpOpFusionModeStandard :: FPOpFusionMode

newtype TargetOptionFlag

Constructors

TargetOptionFlag CUInt

fpOpFusionModeStrict :: FPOpFusionMode

targetOptionFlagPrintMachineCode :: TargetOptionFlag

fpOpFusionModeP :: QuasiQuoter

targetOptionFlagUnsafeFPMath :: TargetOptionFlag

newtype DebugCompressionType

Constructors

DebugCompressionType CUInt

targetOptionFlagNoInfsFPMath :: TargetOptionFlag

debugCompressionTypeNone :: DebugCompressionType

targetOptionFlagNoNaNsFPMath :: TargetOptionFlag

debugCompressionTypeGNU :: DebugCompressionType

newtype TypeKind

Constructors

TypeKind CUInt

targetOptionFlagHonorSignDependentRoundingFPMathOption :: TargetOptionFlag

debugCompressionTypeZ :: DebugCompressionType

typeKindVoid :: TypeKind

targetOptionFlagNoZerosInBSS :: TargetOptionFlag

debugCompressionTypeP :: QuasiQuoter

typeKindHalf :: TypeKind

targetOptionFlagGuaranteedTailCallOpt :: TargetOptionFlag

typeKindFloat :: TypeKind

targetOptionFlagEnableFastISel :: TargetOptionFlag

typeKindDouble :: TypeKind

targetOptionFlagUseInitArray :: TargetOptionFlag

typeKindX86_FP80 :: TypeKind

targetOptionFlagDisableIntegratedAS :: TargetOptionFlag

typeKindFP128 :: TypeKind

targetOptionFlagTrapUnreachable :: TargetOptionFlag

typeKindPPC_FP128 :: TypeKind

targetOptionFlagP :: QuasiQuoter

newtype ParameterAttributeKind

Constructors

ParameterAttributeKind CUInt

typeKindLabel :: TypeKind

parameterAttributeKindAlignment :: ParameterAttributeKind

typeKindInteger :: TypeKind

parameterAttributeKindByVal :: ParameterAttributeKind

newtype FunctionAttributeKind

Constructors

FunctionAttributeKind CUInt

typeKindFunction :: TypeKind

parameterAttributeKindDereferenceable :: ParameterAttributeKind

typeKindStruct :: TypeKind

functionAttributeKindAllocSize :: FunctionAttributeKind

parameterAttributeKindDereferenceableOrNull :: ParameterAttributeKind

typeKindArray :: TypeKind

functionAttributeKindAlwaysInline :: FunctionAttributeKind

newtype FloatSemantics

Constructors

FloatSemantics CUInt

parameterAttributeKindInAlloca :: ParameterAttributeKind

typeKindPointer :: TypeKind

functionAttributeKindArgMemOnly :: FunctionAttributeKind

floatSemanticsIEEEhalf :: FloatSemantics

parameterAttributeKindInReg :: ParameterAttributeKind

typeKindVector :: TypeKind

functionAttributeKindBuiltin :: FunctionAttributeKind

floatSemanticsIEEEsingle :: FloatSemantics

parameterAttributeKindNest :: ParameterAttributeKind

newtype VerifierFailureAction

Constructors

VerifierFailureAction CUInt

typeKindMetadata :: TypeKind

functionAttributeKindCold :: FunctionAttributeKind

floatSemanticsIEEEdouble :: FloatSemantics

parameterAttributeKindNoAlias :: ParameterAttributeKind

typeKindX86_MMX :: TypeKind

functionAttributeKindConvergent :: FunctionAttributeKind

verifierFailureActionAbortProcess :: VerifierFailureAction

floatSemanticsIEEEquad :: FloatSemantics

parameterAttributeKindNoCapture :: ParameterAttributeKind

typeKindToken :: TypeKind

verifierFailureActionPrintMessage :: VerifierFailureAction

functionAttributeKindInaccessibleMemOnly :: FunctionAttributeKind

newtype LibFunc

Constructors

LibFunc CUInt

floatSemanticsPPCDoubleDouble :: FloatSemantics

parameterAttributeKindNonNull :: ParameterAttributeKind

typeKindP :: QuasiQuoter

verifierFailureActionReturnStatus :: VerifierFailureAction

functionAttributeKindInaccessibleMemOrArgMemOnly :: FunctionAttributeKind

libFunc__under_IO_getc :: LibFunc

floatSemanticsx87DoubleExtended :: FloatSemantics

parameterAttributeKindReadNone :: ParameterAttributeKind

verifierFailureActionP :: QuasiQuoter

functionAttributeKindInlineHint :: FunctionAttributeKind

libFunc__under_IO_putc :: LibFunc

floatSemanticsBogus :: FloatSemantics

parameterAttributeKindReadOnly :: ParameterAttributeKind

newtype JITSymbolFlags

Constructors

JITSymbolFlags CUInt

functionAttributeKindJumpTable :: FunctionAttributeKind

libFunc__ZdaPv :: LibFunc

floatSemanticsP :: QuasiQuoter

parameterAttributeKindReturned :: ParameterAttributeKind

jitSymbolFlagsNone :: JITSymbolFlags

functionAttributeKindMinSize :: FunctionAttributeKind

libFunc__ZdaPvRKSt9nothrow_t :: LibFunc

parameterAttributeKindSExt :: ParameterAttributeKind

jitSymbolFlagsWeak :: JITSymbolFlags

functionAttributeKindNaked :: FunctionAttributeKind

libFunc__ZdlPv :: LibFunc

parameterAttributeKindStructRet :: ParameterAttributeKind

jitSymbolFlagsExported :: JITSymbolFlags

functionAttributeKindNoBuiltin :: FunctionAttributeKind

libFunc__ZdlPvRKSt9nothrow_t :: LibFunc

parameterAttributeKindSwiftError :: ParameterAttributeKind

jitSymbolFlagsP :: QuasiQuoter

functionAttributeKindNoDuplicate :: FunctionAttributeKind

libFunc__Znaj :: LibFunc

parameterAttributeKindSwiftSelf :: ParameterAttributeKind

functionAttributeKindNoImplicitFloat :: FunctionAttributeKind

libFunc__ZnajRKSt9nothrow_t :: LibFunc

parameterAttributeKindWriteOnly :: ParameterAttributeKind

functionAttributeKindNoInline :: FunctionAttributeKind

libFunc__Znam :: LibFunc

parameterAttributeKindZExt :: ParameterAttributeKind

functionAttributeKindNoRecurse :: FunctionAttributeKind

libFunc__ZnamRKSt9nothrow_t :: LibFunc

parameterAttributeKindP :: QuasiQuoter

functionAttributeKindNoRedZone :: FunctionAttributeKind

libFunc__Znwj :: LibFunc

functionAttributeKindNoReturn :: FunctionAttributeKind

libFunc__ZnwjRKSt9nothrow_t :: LibFunc

functionAttributeKindNoUnwind :: FunctionAttributeKind

libFunc__Znwm :: LibFunc

functionAttributeKindNonLazyBind :: FunctionAttributeKind

libFunc__ZnwmRKSt9nothrow_t :: LibFunc

functionAttributeKindOptimizeForSize :: FunctionAttributeKind

libFunc__cospi :: LibFunc

functionAttributeKindOptimizeNone :: FunctionAttributeKind

libFunc__cospif :: LibFunc

functionAttributeKindReadNone :: FunctionAttributeKind

libFunc__cxa_atexit :: LibFunc

functionAttributeKindReadOnly :: FunctionAttributeKind

libFunc__cxa_guard_abort :: LibFunc

functionAttributeKindReturnsTwice :: FunctionAttributeKind

libFunc__cxa_guard_acquire :: LibFunc

functionAttributeKindSafeStack :: FunctionAttributeKind

libFunc__cxa_guard_release :: LibFunc

functionAttributeKindSanitizeAddress :: FunctionAttributeKind

libFunc__dunder_isoc99_scanf :: LibFunc

functionAttributeKindSanitizeMemory :: FunctionAttributeKind

libFunc__dunder_isoc99_sscanf :: LibFunc

functionAttributeKindSanitizeThread :: FunctionAttributeKind

libFunc__memcpy_chk :: LibFunc

functionAttributeKindSpeculatable :: FunctionAttributeKind

libFunc__sincospi_stret :: LibFunc

functionAttributeKindStackAlignment :: FunctionAttributeKind

libFunc__sincospif_stret :: LibFunc

functionAttributeKindStackProtect :: FunctionAttributeKind

libFunc__sinpi :: LibFunc

functionAttributeKindStackProtectReq :: FunctionAttributeKind

libFunc__sinpif :: LibFunc

functionAttributeKindStackProtectStrong :: FunctionAttributeKind

libFunc__sqrt_finite :: LibFunc

functionAttributeKindUWTable :: FunctionAttributeKind

libFunc__sqrtf_finite :: LibFunc

functionAttributeKindWriteOnly :: FunctionAttributeKind

libFunc__sqrtl_finite :: LibFunc

functionAttributeKindP :: QuasiQuoter

libFunc__dunder_strdup :: LibFunc

libFunc__dunder_strndup :: LibFunc

libFunc__dunder_strtok_r :: LibFunc

libFunc__abs :: LibFunc

libFunc__access :: LibFunc

libFunc__acos :: LibFunc

libFunc__acosf :: LibFunc

libFunc__acosh :: LibFunc

libFunc__acoshf :: LibFunc

libFunc__acoshl :: LibFunc

libFunc__acosl :: LibFunc

libFunc__asin :: LibFunc

libFunc__asinf :: LibFunc

libFunc__asinh :: LibFunc

libFunc__asinhf :: LibFunc

libFunc__asinhl :: LibFunc

libFunc__asinl :: LibFunc

libFunc__atan :: LibFunc

libFunc__atan2 :: LibFunc

libFunc__atan2f :: LibFunc

libFunc__atan2l :: LibFunc

libFunc__atanf :: LibFunc

libFunc__atanh :: LibFunc

libFunc__atanhf :: LibFunc

libFunc__atanhl :: LibFunc

libFunc__atanl :: LibFunc

libFunc__atof :: LibFunc

libFunc__atoi :: LibFunc

libFunc__atol :: LibFunc

libFunc__atoll :: LibFunc

libFunc__bcmp :: LibFunc

libFunc__bcopy :: LibFunc

libFunc__bzero :: LibFunc

libFunc__calloc :: LibFunc

libFunc__cbrt :: LibFunc

libFunc__cbrtf :: LibFunc

libFunc__cbrtl :: LibFunc

libFunc__ceil :: LibFunc

libFunc__ceilf :: LibFunc

libFunc__ceill :: LibFunc

libFunc__chmod :: LibFunc

libFunc__chown :: LibFunc

libFunc__clearerr :: LibFunc

libFunc__closedir :: LibFunc

libFunc__copysign :: LibFunc

libFunc__copysignf :: LibFunc

libFunc__copysignl :: LibFunc

libFunc__cos :: LibFunc

libFunc__cosf :: LibFunc

libFunc__cosh :: LibFunc

libFunc__coshf :: LibFunc

libFunc__coshl :: LibFunc

libFunc__cosl :: LibFunc

libFunc__ctermid :: LibFunc

libFunc__exp :: LibFunc

libFunc__exp10 :: LibFunc

libFunc__exp10f :: LibFunc

libFunc__exp10l :: LibFunc

libFunc__exp2 :: LibFunc

libFunc__exp2f :: LibFunc

libFunc__exp2l :: LibFunc

libFunc__expf :: LibFunc

libFunc__expl :: LibFunc

libFunc__expm1 :: LibFunc

libFunc__expm1f :: LibFunc

libFunc__expm1l :: LibFunc

libFunc__fabs :: LibFunc

libFunc__fabsf :: LibFunc

libFunc__fabsl :: LibFunc

libFunc__fclose :: LibFunc

libFunc__fdopen :: LibFunc

libFunc__feof :: LibFunc

libFunc__ferror :: LibFunc

libFunc__fflush :: LibFunc

libFunc__ffs :: LibFunc

libFunc__ffsl :: LibFunc

libFunc__ffsll :: LibFunc

libFunc__fgetc :: LibFunc

libFunc__fgetpos :: LibFunc

libFunc__fgets :: LibFunc

libFunc__fileno :: LibFunc

libFunc__fiprintf :: LibFunc

libFunc__flockfile :: LibFunc

libFunc__floor :: LibFunc

libFunc__floorf :: LibFunc

libFunc__floorl :: LibFunc

libFunc__fmax :: LibFunc

libFunc__fmaxf :: LibFunc

libFunc__fmaxl :: LibFunc

libFunc__fmin :: LibFunc

libFunc__fminf :: LibFunc

libFunc__fminl :: LibFunc

libFunc__fmod :: LibFunc

libFunc__fmodf :: LibFunc

libFunc__fmodl :: LibFunc

libFunc__fopen :: LibFunc

libFunc__fopen64 :: LibFunc

libFunc__fprintf :: LibFunc

libFunc__fputc :: LibFunc

libFunc__fputs :: LibFunc

libFunc__fread :: LibFunc

libFunc__free :: LibFunc

libFunc__frexp :: LibFunc

libFunc__frexpf :: LibFunc

libFunc__frexpl :: LibFunc

libFunc__fscanf :: LibFunc

libFunc__fseek :: LibFunc

libFunc__fseeko :: LibFunc

libFunc__fseeko64 :: LibFunc

libFunc__fsetpos :: LibFunc

libFunc__fstat :: LibFunc

libFunc__fstat64 :: LibFunc

libFunc__fstatvfs :: LibFunc

libFunc__fstatvfs64 :: LibFunc

libFunc__ftell :: LibFunc

libFunc__ftello :: LibFunc

libFunc__ftello64 :: LibFunc

libFunc__ftrylockfile :: LibFunc

libFunc__funlockfile :: LibFunc

libFunc__fwrite :: LibFunc

libFunc__getc :: LibFunc

libFunc__getc_unlocked :: LibFunc

libFunc__getchar :: LibFunc

libFunc__getenv :: LibFunc

libFunc__getitimer :: LibFunc

libFunc__getlogin_r :: LibFunc

libFunc__getpwnam :: LibFunc

libFunc__gets :: LibFunc

libFunc__gettimeofday :: LibFunc

libFunc__htonl :: LibFunc

libFunc__htons :: LibFunc

libFunc__iprintf :: LibFunc

libFunc__isascii :: LibFunc

libFunc__isdigit :: LibFunc

libFunc__labs :: LibFunc

libFunc__lchown :: LibFunc

libFunc__ldexp :: LibFunc

libFunc__ldexpf :: LibFunc

libFunc__ldexpl :: LibFunc

libFunc__llabs :: LibFunc

libFunc__log :: LibFunc

libFunc__log10 :: LibFunc

libFunc__log10f :: LibFunc

libFunc__log10l :: LibFunc

libFunc__log1p :: LibFunc

libFunc__log1pf :: LibFunc

libFunc__log1pl :: LibFunc

libFunc__log2 :: LibFunc

libFunc__log2f :: LibFunc

libFunc__log2l :: LibFunc

libFunc__logb :: LibFunc

libFunc__logbf :: LibFunc

libFunc__logbl :: LibFunc

libFunc__logf :: LibFunc

libFunc__logl :: LibFunc

libFunc__lstat :: LibFunc

libFunc__lstat64 :: LibFunc

libFunc__malloc :: LibFunc

libFunc__memalign :: LibFunc

libFunc__memccpy :: LibFunc

libFunc__memchr :: LibFunc

libFunc__memcmp :: LibFunc

libFunc__memcpy :: LibFunc

libFunc__memmove :: LibFunc

libFunc__memrchr :: LibFunc

libFunc__memset :: LibFunc

libFunc__memset_pattern16 :: LibFunc

libFunc__mkdir :: LibFunc

libFunc__mktime :: LibFunc

libFunc__modf :: LibFunc

libFunc__modff :: LibFunc

libFunc__modfl :: LibFunc

libFunc__nearbyint :: LibFunc

libFunc__nearbyintf :: LibFunc

libFunc__nearbyintl :: LibFunc

libFunc__ntohl :: LibFunc

libFunc__ntohs :: LibFunc

libFunc__open :: LibFunc

libFunc__open64 :: LibFunc

libFunc__opendir :: LibFunc

libFunc__pclose :: LibFunc

libFunc__perror :: LibFunc

libFunc__popen :: LibFunc

libFunc__posix_memalign :: LibFunc

libFunc__pow :: LibFunc

libFunc__powf :: LibFunc

libFunc__powl :: LibFunc

libFunc__pread :: LibFunc

libFunc__printf :: LibFunc

libFunc__putc :: LibFunc

libFunc__putchar :: LibFunc

libFunc__puts :: LibFunc

libFunc__pwrite :: LibFunc

libFunc__qsort :: LibFunc

libFunc__read :: LibFunc

libFunc__readlink :: LibFunc

libFunc__realloc :: LibFunc

libFunc__reallocf :: LibFunc

libFunc__realpath :: LibFunc

libFunc__remove :: LibFunc

libFunc__rename :: LibFunc

libFunc__rewind :: LibFunc

libFunc__rint :: LibFunc

libFunc__rintf :: LibFunc

libFunc__rintl :: LibFunc

libFunc__rmdir :: LibFunc

libFunc__round :: LibFunc

libFunc__roundf :: LibFunc

libFunc__roundl :: LibFunc

libFunc__scanf :: LibFunc

libFunc__setbuf :: LibFunc

libFunc__setitimer :: LibFunc

libFunc__setvbuf :: LibFunc

libFunc__sin :: LibFunc

libFunc__sinf :: LibFunc

libFunc__sinh :: LibFunc

libFunc__sinhf :: LibFunc

libFunc__sinhl :: LibFunc

libFunc__sinl :: LibFunc

libFunc__siprintf :: LibFunc

libFunc__snprintf :: LibFunc

libFunc__sprintf :: LibFunc

libFunc__sqrt :: LibFunc

libFunc__sqrtf :: LibFunc

libFunc__sqrtl :: LibFunc

libFunc__sscanf :: LibFunc

libFunc__stat :: LibFunc

libFunc__stat64 :: LibFunc

libFunc__statvfs :: LibFunc

libFunc__statvfs64 :: LibFunc

libFunc__stpcpy :: LibFunc

libFunc__stpncpy :: LibFunc

libFunc__strcasecmp :: LibFunc

libFunc__strcat :: LibFunc

libFunc__strchr :: LibFunc

libFunc__strcmp :: LibFunc

libFunc__strcoll :: LibFunc

libFunc__strcpy :: LibFunc

libFunc__strcspn :: LibFunc

libFunc__strdup :: LibFunc

libFunc__strlen :: LibFunc

libFunc__strncasecmp :: LibFunc

libFunc__strncat :: LibFunc

libFunc__strncmp :: LibFunc

libFunc__strncpy :: LibFunc

libFunc__strndup :: LibFunc

libFunc__strnlen :: LibFunc

libFunc__strpbrk :: LibFunc

libFunc__strrchr :: LibFunc

libFunc__strspn :: LibFunc

libFunc__strstr :: LibFunc

libFunc__strtod :: LibFunc

libFunc__strtof :: LibFunc

libFunc__strtok :: LibFunc

libFunc__strtok_r :: LibFunc

libFunc__strtol :: LibFunc

libFunc__strtold :: LibFunc

libFunc__strtoll :: LibFunc

libFunc__strtoul :: LibFunc

libFunc__strtoull :: LibFunc

libFunc__strxfrm :: LibFunc

libFunc__system :: LibFunc

libFunc__tan :: LibFunc

libFunc__tanf :: LibFunc

libFunc__tanh :: LibFunc

libFunc__tanhf :: LibFunc

libFunc__tanhl :: LibFunc

libFunc__tanl :: LibFunc

libFunc__times :: LibFunc

libFunc__tmpfile :: LibFunc

libFunc__tmpfile64 :: LibFunc

libFunc__toascii :: LibFunc

libFunc__trunc :: LibFunc

libFunc__truncf :: LibFunc

libFunc__truncl :: LibFunc

libFunc__uname :: LibFunc

libFunc__ungetc :: LibFunc

libFunc__unlink :: LibFunc

libFunc__unsetenv :: LibFunc

libFunc__utime :: LibFunc

libFunc__utimes :: LibFunc

libFunc__valloc :: LibFunc

libFunc__vfprintf :: LibFunc

libFunc__vfscanf :: LibFunc

libFunc__vprintf :: LibFunc

libFunc__vscanf :: LibFunc

libFunc__vsnprintf :: LibFunc

libFunc__vsprintf :: LibFunc

libFunc__vsscanf :: LibFunc

libFunc__write :: LibFunc

libFunc__P :: QuasiQuoter

Orphan instances

Data CUInt

Methods

gfoldl :: (forall d b. Data d => c (d -> b) -> d -> c b) -> (forall g. g -> c g) -> CUInt -> c CUInt

gunfold :: (forall b r. Data b => c (b -> r) -> c r) -> (forall r. r -> c r) -> Constr -> c CUInt

toConstr :: CUInt -> Constr

dataTypeOf :: CUInt -> DataType

dataCast1 :: Typeable (* -> *) t => (forall d. Data d => c (t d)) -> Maybe (c CUInt)

dataCast2 :: Typeable (* -> * -> *) t => (forall d e. (Data d, Data e) => c (t d e)) -> Maybe (c CUInt)

gmapT :: (forall b. Data b => b -> b) -> CUInt -> CUInt

gmapQl :: (r -> r' -> r) -> r -> (forall d. Data d => d -> r') -> CUInt -> r

gmapQr :: (r' -> r -> r) -> r -> (forall d. Data d => d -> r') -> CUInt -> r

gmapQ :: (forall d. Data d => d -> u) -> CUInt -> [u]

gmapQi :: Int -> (forall d. Data d => d -> u) -> CUInt -> u

gmapM :: Monad m => (forall d. Data d => d -> m d) -> CUInt -> m CUInt

gmapMp :: MonadPlus m => (forall d. Data d => d -> m d) -> CUInt -> m CUInt

gmapMo :: MonadPlus m => (forall d. Data d => d -> m d) -> CUInt -> m CUInt