LLVM.Internal.FFI.PtrHierarchy

Description

This module defines typeclasses to represent the relationships of an object-oriented inheritance hierarchy

Documentation

class DescendentOf a b where

a class to represent safe casting of pointers to objects of descendant-classes to ancestor-classes.

Minimal complete definition

Nothing

Methods

upCast :: Ptr b -> Ptr a

class ChildOf b c | c -> b

a class to represent direct parent-child relationships

data Value

http://llvm.org/doxygen/classllvm_1_1Value.html

data Constant

http://llvm.org/doxygen/classllvm_1_1Constant.html

data GlobalValue

http://llvm.org/doxygen/classllvm_1_1GlobalValue.html

data GlobalObject

http://llvm.org/doxygen/classllvm_1_1GlobalObject.html

data GlobalVariable

http://llvm.org/doxygen/classllvm_1_1GlobalVariable.html

data GlobalAlias

http://llvm.org/doxygen/classllvm_1_1GlobalAlias.html

data Function

http://llvm.org/doxygen/classllvm_1_1Function.html

data BasicBlock

http://llvm.org/doxygen/classllvm_1_1BasicBlock.html

data Parameter

http://llvm.org/doxygen/classllvm_1_1Argument.html

data Instruction

http://llvm.org/doxygen/classllvm_1_1Instruction.html

data BinaryOperator

http://llvm.org/doxygen/classllvm_1_1BinaryOperator.html

data User

http://llvm.org/doxygen/classllvm_1_1User.html

data MDNode

http://llvm.org/doxygen/classllvm_1_1MDNode.html

data MDTuple

data MDString

http://llvm.org/doxygen/classllvm_1_1MDString.html

data MDValue

http://llvm.org/doxygen/classllvm_1_1ValueAsMetadata.html

data DIExpression

https://llvm.org/doxygen/classllvm_1_1DIExpression.html

data DIGlobalVariableExpression

https://llvm.org/doxygen/classllvm_1_1DIGlobalVariableExpression.html

data DILocation

https://llvm.org/doxygen/classllvm_1_1DILocation.html

data DINode

https://llvm.org/doxygen/classllvm_1_1DINode.html

data DIImportedEntity

https://llvm.org/doxygen/classllvm_1_1DIImportedEntity.html

data DIObjCProperty

https://llvm.org/doxygen/classllvm_1_1DIObjCProperty.html

data DISubrange

https://llvm.org/doxygen/classllvm_1_1DISubrange.html

data DIEnumerator

https://llvm.org/doxygen/classllvm_1_1DIEnumerator.html

data DIVariable

https://llvm.org/doxygen/classllvm_1_1DIVariable.html

data DILocalVariable

https://llvm.org/doxygen/classllvm_1_1DILocalVariable.html

data DIGlobalVariable

https://llvm.org/doxygen/classllvm_1_1DIGlobalVariable.html

data DITemplateParameter

https://llvm.org/doxygen/classllvm_1_1DITemplateParameter.html

data DITemplateTypeParameter

https://llvm.org/doxygen/classllvm_1_1DITemplateTypeParameter.html

data DITemplateValueParameter

https://llvm.org/doxygen/classllvm_1_1DITemplateValueParameter.html

data DIScope

https://llvm.org/doxygen/classllvm_1_1DIScope.html

data DIModule

https://llvm.org/doxygen/classllvm_1_1DIModule.html

data DINamespace

https://llvm.org/doxygen/classllvm_1_1DINamespace.html

data DIFile

https://llvm.org/doxygen/classllvm_1_1DIFile.html

data DICompileUnit

https://llvm.org/doxygen/classllvm_1_1DICompileUnit.html

data DIType

https://llvm.org/doxygen/classllvm_1_1DIType.html

data DIBasicType

https://llvm.org/doxygen/classllvm_1_1DIBasicType.html

data DIDerivedType

https://llvm.org/doxygen/classllvm_1_1DIDerivedType.html

data DISubroutineType

https://llvm.org/doxygen/classllvm_1_1DISubroutineType.html

data DICompositeType

https://llvm.org/doxygen/classllvm_1_1DICompositeType.html

data DILocalScope

https://llvm.org/doxygen/classllvm_1_1DILocalScope.html

data DILexicalBlockBase

https://llvm.org/doxygen/classllvm_1_1DILexicalBlockBase.html

data DILexicalBlock

https://llvm.org/doxygen/classllvm_1_1DILexicalBlock.html

data DILexicalBlockFile

https://llvm.org/doxygen/classllvm_1_1DILexicalBlockFile.html

data DISubprogram

https://llvm.org/doxygen/classllvm_1_1DISubprogram.html

data DIMacroNode

data DIMacro

data DIMacroFile

data NamedMetadata

http://llvm.org/doxygen/classllvm_1_1NamedMDNode.html

data InlineAsm

http://llvm.org/doxygen/classllvm_1_1InlineAsm.html

data Type

http://llvm.org/doxygen/classllvm_1_1Type.html

data Metadata

http://llvm.org/doxygen/classllvm_1_1Metadata.html

data MetadataAsVal

http://www.llvm.org/docs/doxygen/html/classllvm_1_1MetadataAsValue.html

data RawOStream

http://llvm.org/docs/doxygen/html/classllvm_1_1raw__ostream.html

data RawPWriteStream

http://llvm.org/docs/doxygen/html/classllvm_1_1raw__pwrite__stream.html

data StringRef

https://llvm.org/doxygen/classllvm_1_1StringRef.html