LLVM.Internal.FFI.Transforms

Description

Code used with Template Haskell to build the FFI for transform passes.

Documentation

needsTargetMachine :: String -> Bool

does the constructor for this pass require a TargetMachine object

cName :: String -> String

Translate a Haskell name (used in the public Haskell interface, typically not abbreviated) | for a pass into the (sometimes obscure, sometimes abbreviated) name used in the LLVM C interface. | This translation includes, by choice of prefix, whether the C interface implementation is found in | the LLVM distribution (LLVM prefix) or either not available or broken there and so implemented | as part of this Haskell package (LLVM_Hs_ prefix).