LLVM.Internal.Linking

Documentation

loadLibraryPermanently :: Maybe FilePath -> IO Bool

Loads the given dynamic library permanently. If Nothing is given, this will make the symbols from the current process available.

getSymbolAddressInProcess :: MangledSymbol -> IO WordPtr

Get the address of the given symbol in the current process' address space.