LLVM.Internal.ObjectFile

Documentation

newtype ObjectFile

Constructors

ObjectFile (Ptr ObjectFile)

disposeObjectFile :: ObjectFile -> IO ()

Dispose of an ObjectFile.

createObjectFile :: FilePath -> IO ObjectFile

Create a object file which can later be linked with the LinkingLayer.

Note that the file at path should already be a compiled object file i.e a `.o` file. This does *not* compile source files.

withObjectFile :: FilePath -> (ObjectFile -> IO a) -> IO a

bracket-style wrapper for createObjectFile and disposeObjectFile.