LLVM.Internal.OrcJIT

Documentation

newtype MangledSymbol

A mangled symbol which can be used in findSymbol. This can be created using mangleSymbol.

Constructors

MangledSymbol ByteString

newtype ExecutionSession

Constructors

ExecutionSession (Ptr ExecutionSession)

data JITSymbolFlags

Contrary to the C++ interface, we do not store the HasError flag here. Instead decoding a JITSymbol produces a sumtype based on whether that flag is set or not.

Constructors

JITSymbolFlags

Fields

jitSymbolWeak :: !Bool

Is this a weak symbol?

jitSymbolCommon :: !Bool

Is this a common symbol?

jitSymbolAbsolute :: !Bool

Is this an absolute symbol? This will cause LLVM to use absolute relocations for the symbol even in position independent code.

jitSymbolExported :: !Bool

Is this symbol exported?

defaultJITSymbolFlags :: JITSymbolFlags

data JITSymbol

Constructors

JITSymbol

Fields

jitSymbolAddress :: !WordPtr

The address of the symbol. If you’ve looked up a function, you need to cast this to a FunPtr.

jitSymbolFlags :: !JITSymbolFlags

The flags of this symbol.

data JITSymbolError

Constructors

JITSymbolError ShortByteString

newtype SymbolResolver

Specifies how external symbols in a module added to a CompileLayer should be resolved.

Constructors

SymbolResolver (MangledSymbol -> IO (Either JITSymbolError JITSymbol))

withSymbolResolver :: ExecutionSession -> SymbolResolver -> (Ptr SymbolResolver -> IO a) -> IO a

Create a SymbolResolver that can be used with the JIT.

allocWithCleanup :: IORef [IO ()] -> IO a -> (a -> IO ()) -> IO a

Allocate the resource and register it for cleanup.

allocFunPtr :: IORef [IO ()] -> IO (FunPtr a) -> IO (FunPtr a)

Allocate a function pointer and register it for cleanup.

createRegisteredDataLayout :: MonadAnyCont IO m => TargetMachine -> IORef [IO ()] -> m (Ptr DataLayout)

createExecutionSession :: IO ExecutionSession

Create a new ExecutionSession.

disposeExecutionSession :: ExecutionSession -> IO ()

Dispose of an ExecutionSession. This should be called when the ExecutionSession is not needed anymore.

withExecutionSession :: (ExecutionSession -> IO a) -> IO a

bracket-style wrapper around createExecutionSession and disposeExecutionSession.

allocateModuleKey :: ExecutionSession -> IO ModuleKey

Allocate a module key for a new module to add to the JIT.

releaseModuleKey :: ExecutionSession -> ModuleKey -> IO ()

Return a module key to the ExecutionSession so that it can be re-used.

withModuleKey :: ExecutionSession -> (ModuleKey -> IO a) -> IO a

bracket-style wrapper around allocateModuleKey and releaseModuleKey.