LLVM.Internal.OrcJIT.CompileLayer

Documentation

class CompileLayer l where

There are two main types of operations provided by instances of CompileLayer.

You can add / remove modules using addModule / removeModuleSet.

You can search for symbols using findSymbol / findSymbolIn in the previously added modules.

Minimal complete definition

getCompileLayer, getDataLayout, getCleanups

Methods

getCompileLayer :: l -> Ptr CompileLayer

getDataLayout :: l -> Ptr DataLayout

getCleanups :: l -> IORef [IO ()]

mangleSymbol :: CompileLayer l => l -> ShortByteString -> IO MangledSymbol

Mangle a symbol according to the data layout stored in the CompileLayer.

findSymbol :: CompileLayer l => l -> MangledSymbol -> Bool -> IO JITSymbol

findSymbol layer symbol exportedSymbolsOnly searches for symbol in all modules added to layer. If exportedSymbolsOnly is True only exported symbols are searched.

findSymbolIn :: CompileLayer l => l -> ModuleHandle -> MangledSymbol -> Bool -> IO JITSymbol

findSymbolIn layer handle symbol exportedSymbolsOnly searches for symbol in the context of the module represented by handle. If exportedSymbolsOnly is True only exported symbols are searched.

addModule :: CompileLayer l => l -> Module -> SymbolResolver -> IO ModuleHandle

Add a module to the CompileLayer. The SymbolResolver is used to resolve external symbols in the module.

Note: This function consumes the module passed to it and it must not be used after calling this method.

removeModule :: CompileLayer l => l -> ModuleHandle -> IO ()

Remove a previously added module.

withModule :: CompileLayer l => l -> Module -> SymbolResolver -> (ModuleHandle -> IO a) -> IO a

bracket-style wrapper around addModule and removeModule.

Note: This function consumes the module passed to it and it must not be used after calling this method.

disposeCompileLayer :: CompileLayer l => l -> IO ()

Dispose of a CompileLayer. This should called when the CompileLayer is not needed anymore.

data ModuleHandle

Abstract type representing a module in a CompileLayer.