LLVM.Internal.OrcJIT.CompileOnDemandLayer

Documentation

type PartitioningFn = Ptr Function -> IO [Ptr Function]

data JITCompileCallbackManager

This is used by CompileOnDemandLayer to create callback that compile functions when they are called.

Constructors

CallbackMgr !(Ptr JITCompileCallbackManager) !(IO ())

newtype IndirectStubsManagerBuilder

This is used by CompileOnDemandLayer to manage the stubs created for function definitions that have not yet been compiled.

Constructors

StubsMgr (Ptr IndirectStubsManagerBuilder)

data CompileOnDemandLayer baseLayer

Adding a module to a CompileOnDemandLayer creates stubs for its functions definitions. When one of those stubs is called, the corresponding function body is extracted and compiled.

Constructors

CompileOnDemandLayer

Fields

compileLayer :: !(Ptr CompileOnDemandLayer)

dataLayout :: !(Ptr DataLayout)

cleanupActions :: !(IORef [IO ()])

newIndirectStubsManagerBuilder

Arguments

:: ShortByteString
target triple

-> IO IndirectStubsManagerBuilder

Create a new IndirectStubsManagerBuilder.

When the stubs manager is no longer needed, it should be freed using disposeIndirectStubsManagerBuilder.

disposeIndirectStubsManagerBuilder :: IndirectStubsManagerBuilder -> IO ()

Dispose of an IndirectStubsManagerBuilder.

withIndirectStubsManagerBuilder

Arguments

:: ShortByteString
target triple

-> (IndirectStubsManagerBuilder -> IO a)

-> IO a

bracket-style wrapper around newIndirectStubsManagerBuilder and disposeIndirectStubsManagerBuilder.

newJITCompileCallbackManager

Arguments

:: ShortByteString
target triple

-> Maybe (IO ())
called on compilation errors

-> IO JITCompileCallbackManager

Create a new JITCompileCallbackManager.

When the callback manager is no longer needed, it should be freed using disposeJITCompileCallbackManager.

disposeJITCompileCallbackManager :: JITCompileCallbackManager -> IO ()

Dispose of a JITCompileCallbackManager.

withJITCompileCallbackManager

Arguments

:: ShortByteString
target triple

-> Maybe (IO ())
called on compilation errors

-> (JITCompileCallbackManager -> IO a)

-> IO a

Execute a computation using a new JITCompileCallbackManager.

newCompileOnDemandLayer

Arguments

:: CompileLayer l

=> l

-> TargetMachine

-> (Ptr Function -> IO [Ptr Function])
partitioning function

-> JITCompileCallbackManager

-> IndirectStubsManagerBuilder

-> Bool
clone stubs into partitions

-> IO (CompileOnDemandLayer l)

Create a new CompileOnDemandLayer. The partitioning function specifies which functions should be compiled when a function is called.

When the layer is no longer needed, it should be disposed using disposeCompileLayer.

withCompileOnDemandLayer

Arguments

:: CompileLayer l

=> l

-> TargetMachine

-> (Ptr Function -> IO [Ptr Function])
partitioning function

-> JITCompileCallbackManager

-> IndirectStubsManagerBuilder

-> Bool
clone stubs into partitions

-> (CompileOnDemandLayer l -> IO a)

-> IO a

bracket-style wrapper around newCompileOnDemandLayer and disposeCompileLayer.

Orphan instances

MonadIO m => EncodeM m PartitioningFn (IORef [IO ()] -> IO (FunPtr PartitioningFn))

Methods

encodeM :: PartitioningFn -> m (IORef [IO ()] -> IO (FunPtr PartitioningFn))

(MonadIO m, MonadAnyCont IO m) => EncodeM m (Maybe (IO ())) (TargetAddress, IO ())

Methods

encodeM :: Maybe (IO ()) -> m (TargetAddress, IO ())