LLVM.Module

Description

A Module holds a C++ LLVM IR module. Modules may be converted to or from strings or Haskell ASTs, or added to an ExecutionEngine and so JIT compiled to get function pointers.

Documentation

data Module

http://llvm.org/doxygen/classllvm_1_1Module.html

newtype File

A newtype to distinguish strings used for paths from other strings

Constructors

File FilePath

withModuleFromAST :: Context -> Module -> (Module -> IO a) -> IO a

Execute a function after encoding the module in LLVM’s internal representation. May throw EncodeException.

moduleAST :: Module -> IO Module

Get an LLVM.AST.Module from a LLVM.Module - i.e. raise C++ objects into an Haskell AST.

withModuleFromLLVMAssembly :: LLVMAssemblyInput s => Context -> s -> (Module -> IO a) -> IO a

parse Module from LLVM assembly. May throw ParseFailureException.

moduleLLVMAssembly :: Module -> IO ByteString

generate LLVM assembly from a Module

writeLLVMAssemblyToFile :: File -> Module -> IO ()

write LLVM assembly for a Module to a file

withModuleFromBitcode :: BitcodeInput b => Context -> b -> (Module -> IO a) -> IO a

parse Module from LLVM bitcode. May throw ParseFailureException.

moduleBitcode :: Module -> IO ByteString

generate LLVM bitcode from a Module

writeBitcodeToFile :: File -> Module -> IO ()

write LLVM bitcode from a Module into a file

moduleTargetAssembly :: TargetMachine -> Module -> IO ByteString

produce target-specific assembly as a ByteString

writeTargetAssemblyToFile :: TargetMachine -> File -> Module -> IO ()

write target-specific assembly directly into a file

moduleObject :: TargetMachine -> Module -> IO ByteString

produce target-specific object code as a ByteString

writeObjectToFile :: TargetMachine -> File -> Module -> IO ()

write target-specific object code directly into a file

linkModules

Arguments

:: Module
The module into which to link

-> Module
The module to link into the other (this module is destroyed)

-> IO ()

link LLVM modules - move or copy parts of a source module into a destination module. Note that this operation is not commutative - not only concretely (e.g. the destination module is modified, becoming the result) but abstractly (e.g. unused private globals in the source module do not appear in the result, but similar globals in the destination remain). The source module is destroyed. May throw a LinkException.