LLVM.OrcJIT.LinkingLayer

Documentation

class LinkingLayer l where

After a CompileLayer has compiled the modules to object code, it passes the resulting object files to a LinkingLayer.

Methods

getLinkingLayer :: l -> Ptr LinkingLayer

getCleanups :: l -> IORef [IO ()]

disposeLinkingLayer :: LinkingLayer l => l -> IO ()

Dispose of a LinkingLayer.

data ObjectLinkingLayer

Bare bones implementation of a LinkingLayer.

Constructors

ObjectLinkingLayer

Fields

linkingLayer :: !(Ptr ObjectLinkingLayer)

cleanupActions :: !(IORef [IO ()])

newObjectLinkingLayer :: ExecutionSession -> (ModuleKey -> IO (Ptr SymbolResolver)) -> IO ObjectLinkingLayer

Create a new ObjectLinkingLayer. This should be disposed using disposeLinkingLayer when it is no longer needed.

withObjectLinkingLayer :: ExecutionSession -> (ModuleKey -> IO (Ptr SymbolResolver)) -> (ObjectLinkingLayer -> IO a) -> IO a

bracket-style wrapper around newObjectLinkingLayer and disposeLinkingLayer.

addObjectFile :: LinkingLayer l => l -> ModuleKey -> ObjectFile -> IO ()

Add an object file to the LinkingLayer.

findSymbol :: LinkingLayer l => l -> ShortByteString -> Bool -> IO (Either JITSymbolError JITSymbol)

findSymbol layer symbol exportedSymbolsOnly searches for symbol in all modules added to layer. If exportedSymbolsOnly is True only exported symbols are searched.

findSymbolIn :: LinkingLayer l => l -> ModuleKey -> ShortByteString -> Bool -> IO (Either JITSymbolError JITSymbol)

findSymbolIn layer handle symbol exportedSymbolsOnly searches for symbol in the context of the module represented by handle. If exportedSymbolsOnly is True only exported symbols are searched.