LLVM.Target

Description

A Target describes that for which code may be intended. Targets are used both during actual | lowering of LLVM IR to machine code and by some optimization passes which use the target to | judge costs.

Documentation

lookupTarget

Arguments

:: Maybe ShortByteString
arch

-> ShortByteString
"triple" - e.g. x86_64-unknown-linux-gnu

-> IO (Target, ShortByteString)

Find a Target given an architecture and/or a "triple". | http://llvm.org/doxygen/structllvm_1_1TargetRegistry.html#a3105b45e546c9cc3cf78d0f2ec18ad89 | Be sure to run either initializeAllTargets or initializeNativeTarget before expecting this to succeed, depending on what target(s) you want to use. May throw LookupTargetException if no target is found.

data TargetOptions

http://llvm.org/doxygen/classllvm_1_1TargetOptions.html

data Target

http://llvm.org/doxygen/classllvm_1_1Target.html

data TargetMachine

http://llvm.org/doxygen/classllvm_1_1TargetMachine.html

data TargetLowering

http://llvm.org/doxygen/classllvm_1_1TargetLowering.html

newtype CPUFeature

e.g. an instruction set extension

Constructors

CPUFeature ByteString

withTargetOptions :: (TargetOptions -> IO a) -> IO a

bracket creation and destruction of a TargetOptions object

peekTargetOptions :: TargetOptions -> IO Options

get all target options

pokeTargetOptions :: Options -> TargetOptions -> IO ()

set all target options

withTargetMachine

Arguments

:: Target

-> ShortByteString
triple

-> ByteString
cpu

-> Map CPUFeature Bool
features

-> TargetOptions

-> Model

-> Model

-> Level

-> (TargetMachine -> IO a)

-> IO a

bracket creation and destruction of a TargetMachine

withHostTargetMachine :: (TargetMachine -> IO a) -> IO a

Bracket creation and destruction of a TargetMachine configured for the host

targetMachineOptions :: TargetMachine -> IO TargetOptions

getTargetLowering :: TargetMachine -> IO TargetLowering

get the TargetLowering of a TargetMachine

getTargetMachineTriple :: TargetMachine -> IO ShortByteString

the target triple corresponding to the target machine

getDefaultTargetTriple :: IO ShortByteString

the default target triple that LLVM has been configured to produce code for

getProcessTargetTriple :: IO ShortByteString

a target triple suitable for loading code into the current process

getHostCPUName :: IO ByteString

the LLVM name for the host CPU

getHostCPUFeatures :: IO (Map CPUFeature Bool)

a space-separated list of LLVM feature names supported by the host CPU

getTargetMachineDataLayout :: TargetMachine -> IO DataLayout

DataLayout to use for the given TargetMachine

initializeNativeTarget :: IO ()

Initialize the native target. This function is called automatically in these Haskell bindings when creating an ExecutionEngine which will require it, and so it should not be necessary to call it separately.

initializeAllTargets :: IO ()

Initialize all targets so they can be found by lookupTarget

data TargetLibraryInfo

http://llvm.org/docs/doxygen/html/classllvm_1_1TargetLibraryInfo.html

getLibraryFunction :: TargetLibraryInfo -> ShortByteString -> IO (Maybe LibraryFunction)

Look up a LibraryFunction by its standard name

getLibraryFunctionName :: TargetLibraryInfo -> LibraryFunction -> IO ShortByteString

Get a the current name to be emitted for a LibraryFunction

setLibraryFunctionAvailableWithName

Arguments

:: TargetLibraryInfo

-> LibraryFunction

-> ShortByteString
The function name to be emitted

-> IO ()

Set the name of the function on the target platform that corresponds to funcName

withTargetLibraryInfo

Arguments

:: ShortByteString
triple

-> (TargetLibraryInfo -> IO a)

-> IO a

look up information about the library functions available on a given platform