LLVM.Exception

Description

This module lists all of the exceptions thrown by 'llvm-hs' itself. Note that other exceptions can potentially be thrown by the underlying libraries, e.g., for functions doing file IO.

Documentation

data EncodeException

Indicates an error during the translation of the AST provided by 'llvm-hs-pure' to LLVM’s internal representation.

Constructors

EncodeException !String

data ParseFailureException

Indicates an error during the parsing of a module. This is used for errors encountered when parsing LLVM’s human readable assembly format and when parsing the binary bitcode format.

Constructors

ParseFailureException !String

data LinkException

Indicates an error during the linking of two modules.

Constructors

LinkException !String

data FdStreamException

Indicates an error during the creation of a raw_fd_ostream. This could be caused by a nonexisting file path.

Constructors

FdStreamException !String

data TargetMachineEmitException

Indicates an error during a call to targetMachineEmit.

Constructors

TargetMachineEmitException !String

data LookupTargetException

Indicates a failure to find the target.

Constructors

LookupTargetException !String

data VerifyException

Indicates an error during the verification of a module.

Constructors

VerifyException !String