LLVM.Internal.FFI.PtrHierarchy

Description

This module defines typeclasses to represent the relationships of an object-oriented inheritance hierarchy

Documentation

class DescendentOf a b where

a class to represent safe casting of pointers to objects of descendant-classes to ancestor-classes.

Methods

upCast :: Ptr b -> Ptr a

class ChildOf b c | c -> b

a class to represent direct parent-child relationships

data Value

http://llvm.org/doxygen/classllvm_1_1Value.html

data Constant

http://llvm.org/doxygen/classllvm_1_1Constant.html

data GlobalValue

http://llvm.org/doxygen/classllvm_1_1GlobalValue.html

data GlobalObject

http://llvm.org/doxygen/classllvm_1_1GlobalObject.html

data GlobalVariable

http://llvm.org/doxygen/classllvm_1_1GlobalVariable.html

data GlobalAlias

http://llvm.org/doxygen/classllvm_1_1GlobalAlias.html

data Function

http://llvm.org/doxygen/classllvm_1_1Function.html

data BasicBlock

http://llvm.org/doxygen/classllvm_1_1BasicBlock.html

data Parameter

http://llvm.org/doxygen/classllvm_1_1Argument.html

data Instruction

http://llvm.org/doxygen/classllvm_1_1Instruction.html

data BinaryOperator

http://llvm.org/doxygen/classllvm_1_1BinaryOperator.html

data User

http://llvm.org/doxygen/classllvm_1_1User.html

data MDNode

http://llvm.org/doxygen/classllvm_1_1MDNode.html

data MDString

http://llvm.org/doxygen/classllvm_1_1MDString.html

data MDValue

http://llvm.org/doxygen/classllvm_1_1ValueAsMetadata.html

data NamedMetadata

http://llvm.org/doxygen/classllvm_1_1NamedMDNode.html

data InlineAsm

http://llvm.org/doxygen/classllvm_1_1InlineAsm.html

data Type

http://llvm.org/doxygen/classllvm_1_1Type.html

data Metadata

http://llvm.org/doxygen/classllvm_1_1Metadata.html

data MetadataAsVal

http://www.llvm.org/docs/doxygen/html/classllvm_1_1MetadataAsValue.html

data RawOStream

http://llvm.org/docs/doxygen/html/classllvm_1_1raw__ostream.html

data RawPWriteStream

http://llvm.org/docs/doxygen/html/classllvm_1_1raw__pwrite__stream.html