LLVM.Internal.Module

Description

This Haskell module is for/of functions for handling LLVM modules.

Documentation

newtype Module

http://llvm.org/doxygen/classllvm_1_1Module.html

Constructors

Module (IORef (Ptr Module))

newModule :: Ptr Module -> IO Module

readModule :: MonadIO m => Module -> m (Ptr Module)

deleteModule :: Module -> IO ()

Signal that a module does no longer exist and thus must not be disposed. It is the responsibility of the caller to ensure that the module has been disposed. If you use only the functions provided by llvm-hs you should never call this yourself.

newtype File

A newtype to distinguish strings used for paths from other strings

Constructors

File FilePath

linkModules

Arguments

:: Module
The module into which to link

-> Module
The module to link into the other (this module is destroyed)

-> IO ()

link LLVM modules - move or copy parts of a source module into a destination module. Note that this operation is not commutative - not only concretely (e.g. the destination module is modified, becoming the result) but abstractly (e.g. unused private globals in the source module do not appear in the result, but similar globals in the destination remain). The source module is destroyed. May throw a LinkException.

class LLVMAssemblyInput s where

Minimal complete definition

llvmAssemblyMemoryBuffer

Methods

llvmAssemblyMemoryBuffer :: (MonadThrow m, MonadIO m, MonadAnyCont IO m) => s -> m (OwnerTransfered (Ptr MemoryBuffer))

withModuleFromLLVMAssembly :: LLVMAssemblyInput s => Context -> s -> (Module -> IO a) -> IO a

parse Module from LLVM assembly. May throw ParseFailureException.

moduleLLVMAssembly :: Module -> IO ByteString

generate LLVM assembly from a Module

writeLLVMAssemblyToFile :: File -> Module -> IO ()

write LLVM assembly for a Module to a file

class BitcodeInput b where

Minimal complete definition

bitcodeMemoryBuffer

Methods

bitcodeMemoryBuffer :: (MonadThrow m, MonadIO m, MonadAnyCont IO m) => b -> m (Ptr MemoryBuffer)

withModuleFromBitcode :: BitcodeInput b => Context -> b -> (Module -> IO a) -> IO a

parse Module from LLVM bitcode. May throw ParseFailureException.

moduleBitcode :: Module -> IO ByteString

generate LLVM bitcode from a Module

writeBitcodeToFile :: File -> Module -> IO ()

write LLVM bitcode from a Module into a file

targetMachineEmit :: CodeGenFileType -> TargetMachine -> Module -> Ptr RawPWriteStream -> IO ()

May throw TargetMachineEmitException.

emitToFile :: CodeGenFileType -> TargetMachine -> File -> Module -> IO ()

May throw FdStreamException and TargetMachineEmitException.

emitToByteString :: CodeGenFileType -> TargetMachine -> Module -> IO ByteString

May throw TargetMachineEmitException.

writeTargetAssemblyToFile :: TargetMachine -> File -> Module -> IO ()

write target-specific assembly directly into a file

moduleTargetAssembly :: TargetMachine -> Module -> IO ByteString

produce target-specific assembly as a ByteString

moduleObject :: TargetMachine -> Module -> IO ByteString

produce target-specific object code as a ByteString

writeObjectToFile :: TargetMachine -> File -> Module -> IO ()

write target-specific object code directly into a file

setTargetTriple :: Ptr Module -> ShortByteString -> EncodeAST ()

getTargetTriple :: Ptr Module -> IO (Maybe ShortByteString)

setDataLayout :: Ptr Module -> DataLayout -> EncodeAST ()

getDataLayout :: Ptr Module -> IO (Maybe DataLayout)

withModuleFromAST :: Context -> Module -> (Module -> IO a) -> IO a

Execute a function after encoding the module in LLVM’s internal representation. May throw EncodeException.

decodeGlobalVariables :: Ptr Module -> DecodeAST (DecodeAST [Global])

decodeGlobalAliases :: Ptr Module -> DecodeAST (DecodeAST [Global])

decodeFunctions :: Ptr Module -> DecodeAST (DecodeAST [Global])

decodeNamedMetadataDefinitions :: Ptr Module -> DecodeAST [Definition]

moduleAST :: Module -> IO Module

Get an LLVM.AST.Module from a LLVM.Module - i.e. raise C++ objects into an Haskell AST.