LLVM.Internal.OrcJIT

Documentation

newtype MangledSymbol

A mangled symbol which can be used in findSymbol. This can be created using mangleSymbol.

Constructors

MangledSymbol ByteString

data JITSymbolFlags

Constructors

JITSymbolFlags

Fields

jitSymbolWeak :: !Bool

Is this a weak symbol?

jitSymbolExported :: !Bool

Is this symbol exported?

data JITSymbol

Constructors

JITSymbol

Fields

jitSymbolAddress :: !WordPtr

The address of the symbol. If you’ve looked up a function, you need to cast this to a FunPtr.

jitSymbolFlags :: !JITSymbolFlags

The flags of this symbol.

type SymbolResolverFn = MangledSymbol -> IO JITSymbol

data SymbolResolver

Specifies how external symbols in a module added to a CompielLayer should be resolved.

Constructors

SymbolResolver

Fields

dylibResolver :: !SymbolResolverFn

This is used to find symbols in the same logical dynamic library as the module referencing them.

externalResolver :: !SymbolResolverFn

When dylibResolver fails to resolve a symbol, externalResolver is used as a fallback to find external symbols.

class LinkingLayer l where

After a CompileLayer has compiled the modules to object code, it passes the resulting object files to a LinkingLayer.

Minimal complete definition

getLinkingLayer

Methods

getLinkingLayer :: l -> Ptr LinkingLayer

newtype ObjectLinkingLayer

Bare bones implementation of a LinkingLayer.

Constructors

ObjectLinkingLayer (Ptr ObjectLinkingLayer)

allocFunPtr :: IORef [IO ()] -> IO (FunPtr a) -> IO (FunPtr a)

allocate a function pointer and register it for cleanup

disposeLinkingLayer :: LinkingLayer l => l -> IO ()

Dispose of a LinkingLayer.

newObjectLinkingLayer :: IO ObjectLinkingLayer

Create a new ObjectLinkingLayer. This should be disposed using disposeLinkingLayer when it is no longer needed.

withObjectLinkingLayer :: (ObjectLinkingLayer -> IO a) -> IO a

bracket-style wrapper around newObjectLinkingLayer and disposeLinkingLayer.

createRegisteredDataLayout :: MonadAnyCont IO m => TargetMachine -> IORef [IO ()] -> m (Ptr DataLayout)