LLVM.Internal.OrcJIT.IRCompileLayer

Documentation

data IRCompileLayer linkingLayer

IRCompileLayer compiles modules immediately when they are added. It parametrized by a LinkingLayer which handles linking of the generated object files.

Constructors

IRCompileLayer

Fields

compileLayer :: !(Ptr IRCompileLayer)

dataLayout :: !(Ptr DataLayout)

cleanupActions :: !(IORef [IO ()])

newIRCompileLayer :: LinkingLayer l => l -> TargetMachine -> IO (IRCompileLayer l)

Create a new IRCompileLayer.

When the layer is no longer needed, it should be disposed using 'disposeCompileLayer.

withIRCompileLayer :: LinkingLayer l => l -> TargetMachine -> (IRCompileLayer l -> IO a) -> IO a

bracket-style wrapper around newIRCompileLayer and disposeCompileLayer.