LLVM.Internal.OrcJIT.IRTransformLayer

Documentation

data IRTransformLayer baseLayer

IRTransformLayer allows transforming modules before handing off compilation to the underlying CompileLayer.

Constructors

IRTransformLayer

Fields

compileLayer :: !(Ptr IRTransformLayer)

dataLayout :: !(Ptr DataLayout)

cleanupActions :: !(IORef [IO ()])

newIRTransformLayer

Arguments

:: CompileLayer l

=> l

-> TargetMachine

-> (Ptr Module -> IO (Ptr Module))
module transformation

-> IO (IRTransformLayer l)

Create a new IRTransformLayer.

When the layer is no longer needed, it should be disposed using disposeCompileLayer.

withIRTransformLayer

Arguments

:: CompileLayer l

=> l

-> TargetMachine

-> (Ptr Module -> IO (Ptr Module))
module transformation

-> (IRTransformLayer l -> IO a)

-> IO a

bracket-style wrapper around newIRTransformLayer and disposeCompileLayer.