LLVM.Internal.Target

Documentation

newtype Target

http://llvm.org/doxygen/classllvm_1_1Target.html

Constructors

Target (Ptr Target)

newtype CPUFeature

e.g. an instruction set extension

Constructors

CPUFeature ByteString

lookupTarget

Arguments

:: Maybe ShortByteString
arch

-> ShortByteString
"triple" - e.g. x86_64-unknown-linux-gnu

-> IO (Target, ShortByteString)

Find a Target given an architecture and/or a "triple". | http://llvm.org/doxygen/structllvm_1_1TargetRegistry.html#a3105b45e546c9cc3cf78d0f2ec18ad89 | Be sure to run either initializeAllTargets or initializeNativeTarget before expecting this to succeed, depending on what target(s) you want to use. May throw LookupTargetException if no target is found.

newtype TargetOptions

http://llvm.org/doxygen/classllvm_1_1TargetOptions.html

Constructors

TargetOptions (Ptr TargetOptions)

newtype MCTargetOptions

Constructors

MCTargetOptions (Ptr MCTargetOptions)

withTargetOptions :: (TargetOptions -> IO a) -> IO a

bracket creation and destruction of a TargetOptions object

pokeTargetOptions :: Options -> TargetOptions -> IO ()

set all target options

pokeMachineCodeOptions :: MachineCodeOptions -> MCTargetOptions -> IO ()

peekTargetOptions :: TargetOptions -> IO Options

get all target options

peekMachineCodeOptions :: MCTargetOptions -> IO MachineCodeOptions

get all machine code options

newtype TargetMachine

http://llvm.org/doxygen/classllvm_1_1TargetMachine.html

Constructors

TargetMachine (Ptr TargetMachine)

withTargetMachine

Arguments

:: Target

-> ShortByteString
triple

-> ByteString
cpu

-> Map CPUFeature Bool
features

-> TargetOptions

-> Model

-> Model

-> Level

-> (TargetMachine -> IO a)

-> IO a

bracket creation and destruction of a TargetMachine

targetMachineOptions :: TargetMachine -> IO TargetOptions

machineCodeOptions :: TargetOptions -> IO MCTargetOptions

newtype TargetLowering

http://llvm.org/doxygen/classllvm_1_1TargetLowering.html

Constructors

TargetLowering (Ptr TargetLowering)

getTargetLowering :: TargetMachine -> IO TargetLowering

get the TargetLowering of a TargetMachine

initializeNativeTarget :: IO ()

Initialize the native target. This function is called automatically in these Haskell bindings when creating an ExecutionEngine which will require it, and so it should not be necessary to call it separately.

getTargetMachineTriple :: TargetMachine -> IO ShortByteString

the target triple corresponding to the target machine

getDefaultTargetTriple :: IO ShortByteString

the default target triple that LLVM has been configured to produce code for

getProcessTargetTriple :: IO ShortByteString

a target triple suitable for loading code into the current process

getHostCPUName :: IO ByteString

the LLVM name for the host CPU

getHostCPUFeatures :: IO (Map CPUFeature Bool)

a space-separated list of LLVM feature names supported by the host CPU

getTargetMachineDataLayout :: TargetMachine -> IO DataLayout

DataLayout to use for the given TargetMachine

initializeAllTargets :: IO ()

Initialize all targets so they can be found by lookupTarget

withHostTargetMachine :: (TargetMachine -> IO a) -> IO a

Bracket creation and destruction of a TargetMachine configured for the host

newtype TargetLibraryInfo

http://llvm.org/docs/doxygen/html/classllvm_1_1TargetLibraryInfo.html

Constructors

TargetLibraryInfo (Ptr TargetLibraryInfo)

getLibraryFunction :: TargetLibraryInfo -> ShortByteString -> IO (Maybe LibraryFunction)

Look up a LibraryFunction by its standard name

getLibraryFunctionName :: TargetLibraryInfo -> LibraryFunction -> IO ShortByteString

Get a the current name to be emitted for a LibraryFunction

setLibraryFunctionAvailableWithName

Arguments

:: TargetLibraryInfo

-> LibraryFunction

-> ShortByteString
The function name to be emitted

-> IO ()

Set the name of the function on the target platform that corresponds to funcName

withTargetLibraryInfo

Arguments

:: ShortByteString
triple

-> (TargetLibraryInfo -> IO a)

-> IO a

look up information about the library functions available on a given platform

Orphan instances

Monad m => DecodeM m Model RelocModel

Methods

decodeM :: RelocModel -> m Model

Monad m => DecodeM m Model CodeModel

Methods

decodeM :: CodeModel -> m Model

Monad m => DecodeM m Level CodeGenOptLevel

Methods

decodeM :: CodeGenOptLevel -> m Level

Monad m => DecodeM m FloatABI FloatABIType

Methods

decodeM :: FloatABIType -> m FloatABI

Monad m => DecodeM m FloatingPointOperationFusionMode FPOpFusionMode

Methods

decodeM :: FPOpFusionMode -> m FloatingPointOperationFusionMode

Monad m => DecodeM m DebugCompressionType DebugCompressionType

Methods

decodeM :: DebugCompressionType -> m DebugCompressionType

Monad m => DecodeM m ThreadModel ThreadModel

Methods

decodeM :: ThreadModel -> m ThreadModel

Monad m => DecodeM m EABIVersion EABI

Methods

decodeM :: EABI -> m EABIVersion

Monad m => DecodeM m DebuggerKind DebuggerKind

Methods

decodeM :: DebuggerKind -> m DebuggerKind

Monad m => DecodeM m FloatingPointDenormalMode FPDenormalMode

Methods

decodeM :: FPDenormalMode -> m FloatingPointDenormalMode

Monad m => DecodeM m ExceptionHandling ExceptionHandling

Methods

decodeM :: ExceptionHandling -> m ExceptionHandling

Monad m => EncodeM m Model RelocModel

Methods

encodeM :: Model -> m RelocModel

Monad m => EncodeM m Model CodeModel

Methods

encodeM :: Model -> m CodeModel

Monad m => EncodeM m Level CodeGenOptLevel

Methods

encodeM :: Level -> m CodeGenOptLevel

Monad m => EncodeM m FloatABI FloatABIType

Methods

encodeM :: FloatABI -> m FloatABIType

Monad m => EncodeM m FloatingPointOperationFusionMode FPOpFusionMode

Methods

encodeM :: FloatingPointOperationFusionMode -> m FPOpFusionMode

Monad m => EncodeM m DebugCompressionType DebugCompressionType

Methods

encodeM :: DebugCompressionType -> m DebugCompressionType

Monad m => EncodeM m ThreadModel ThreadModel

Methods

encodeM :: ThreadModel -> m ThreadModel

Monad m => EncodeM m EABIVersion EABI

Methods

encodeM :: EABIVersion -> m EABI

Monad m => EncodeM m DebuggerKind DebuggerKind

Methods

encodeM :: DebuggerKind -> m DebuggerKind

Monad m => EncodeM m FloatingPointDenormalMode FPDenormalMode

Methods

encodeM :: FloatingPointDenormalMode -> m FPDenormalMode

Monad m => EncodeM m ExceptionHandling ExceptionHandling

Methods

encodeM :: ExceptionHandling -> m ExceptionHandling