LLVM.OrcJIT

CompileLayer

class CompileLayer l

There are two main types of operations provided by instances of CompileLayer.

You can add / remove modules using addModule / removeModuleSet.

You can search for symbols using findSymbol / findSymbolIn in the previously added modules.

Minimal complete definition

getCompileLayer, getDataLayout, getCleanups

Add/remove modules

data ModuleHandle

Abstract type representing a module in a CompileLayer.

addModule :: CompileLayer l => l -> Module -> SymbolResolver -> IO ModuleHandle

Add a module to the CompileLayer. The SymbolResolver is used to resolve external symbols in the module.

Note: This function consumes the module passed to it and it must not be used after calling this method.

removeModule :: CompileLayer l => l -> ModuleHandle -> IO ()

Remove a previously added module.

withModule :: CompileLayer l => l -> Module -> SymbolResolver -> (ModuleHandle -> IO a) -> IO a

bracket-style wrapper around addModule and removeModule.

Note: This function consumes the module passed to it and it must not be used after calling this method.

Search for symbols

findSymbol :: CompileLayer l => l -> MangledSymbol -> Bool -> IO JITSymbol

findSymbol layer symbol exportedSymbolsOnly searches for symbol in all modules added to layer. If exportedSymbolsOnly is True only exported symbols are searched.

findSymbolIn :: CompileLayer l => l -> ModuleHandle -> MangledSymbol -> Bool -> IO JITSymbol

findSymbolIn layer handle symbol exportedSymbolsOnly searches for symbol in the context of the module represented by handle. If exportedSymbolsOnly is True only exported symbols are searched.

data JITSymbol

Constructors

JITSymbol

Fields

jitSymbolAddress :: !WordPtr

The address of the symbol. If you’ve looked up a function, you need to cast this to a FunPtr.

jitSymbolFlags :: !JITSymbolFlags

The flags of this symbol.

data JITSymbolFlags

Constructors

JITSymbolFlags

Fields

jitSymbolWeak :: !Bool

Is this a weak symbol?

jitSymbolExported :: !Bool

Is this symbol exported?

data SymbolResolver

Specifies how external symbols in a module added to a CompielLayer should be resolved.

Constructors

SymbolResolver

Fields

dylibResolver :: !SymbolResolverFn

This is used to find symbols in the same logical dynamic library as the module referencing them.

externalResolver :: !SymbolResolverFn

When dylibResolver fails to resolve a symbol, externalResolver is used as a fallback to find external symbols.

Symbol mangling

data MangledSymbol

A mangled symbol which can be used in findSymbol. This can be created using mangleSymbol.

mangleSymbol :: CompileLayer l => l -> ShortByteString -> IO MangledSymbol

Mangle a symbol according to the data layout stored in the CompileLayer.

IRCompileLayer

data IRCompileLayer linkingLayer

IRCompileLayer compiles modules immediately when they are added. It parametrized by a LinkingLayer which handles linking of the generated object files.

newIRCompileLayer :: LinkingLayer l => l -> TargetMachine -> IO (IRCompileLayer l)

Create a new IRCompileLayer.

When the layer is no longer needed, it should be disposed using 'disposeCompileLayer.

withIRCompileLayer :: LinkingLayer l => l -> TargetMachine -> (IRCompileLayer l -> IO a) -> IO a

bracket-style wrapper around newIRCompileLayer and disposeCompileLayer.

CompileOnDemandLayer

data CompileOnDemandLayer baseLayer

Adding a module to a CompileOnDemandLayer creates stubs for its functions definitions. When one of those stubs is called, the corresponding function body is extracted and compiled.

newCompileOnDemandLayer

Arguments

:: CompileLayer l

=> l

-> TargetMachine

-> (Ptr Function -> IO [Ptr Function])
partitioning function

-> JITCompileCallbackManager

-> IndirectStubsManagerBuilder

-> Bool
clone stubs into partitions

-> IO (CompileOnDemandLayer l)

Create a new CompileOnDemandLayer. The partitioning function specifies which functions should be compiled when a function is called.

When the layer is no longer needed, it should be disposed using disposeCompileLayer.

withCompileOnDemandLayer

Arguments

:: CompileLayer l

=> l

-> TargetMachine

-> (Ptr Function -> IO [Ptr Function])
partitioning function

-> JITCompileCallbackManager

-> IndirectStubsManagerBuilder

-> Bool
clone stubs into partitions

-> (CompileOnDemandLayer l -> IO a)

-> IO a

bracket-style wrapper around newCompileOnDemandLayer and disposeCompileLayer.

IRTRansformLayer

data IRTransformLayer baseLayer

IRTransformLayer allows transforming modules before handing off compilation to the underlying CompileLayer.

newIRTransformLayer

Arguments

:: CompileLayer l

=> l

-> TargetMachine

-> (Ptr Module -> IO (Ptr Module))
module transformation

-> IO (IRTransformLayer l)

Create a new IRTransformLayer.

When the layer is no longer needed, it should be disposed using disposeCompileLayer.

withIRTransformLayer

Arguments

:: CompileLayer l

=> l

-> TargetMachine

-> (Ptr Module -> IO (Ptr Module))
module transformation

-> (IRTransformLayer l -> IO a)

-> IO a

bracket-style wrapper around newIRTransformLayer and disposeCompileLayer.

Dispose of compile layers

disposeCompileLayer :: CompileLayer l => l -> IO ()

Dispose of a CompileLayer. This should called when the CompileLayer is not needed anymore.

LinkingLayer

class LinkingLayer l

After a CompileLayer has compiled the modules to object code, it passes the resulting object files to a LinkingLayer.

Minimal complete definition

getLinkingLayer

Create linking layers

data ObjectLinkingLayer

Bare bones implementation of a LinkingLayer.

newObjectLinkingLayer :: IO ObjectLinkingLayer

Create a new ObjectLinkingLayer. This should be disposed using disposeLinkingLayer when it is no longer needed.

withObjectLinkingLayer :: (ObjectLinkingLayer -> IO a) -> IO a

bracket-style wrapper around newObjectLinkingLayer and disposeLinkingLayer.

Dispose of linking layers

disposeLinkingLayer :: LinkingLayer l => l -> IO ()

Dispose of a LinkingLayer.

JITCompileCallbackManager

data JITCompileCallbackManager

This is used by CompileOnDemandLayer to create callback that compile functions when they are called.

newJITCompileCallbackManager

Arguments

:: ShortByteString
target triple

-> Maybe (IO ())
called on compilation errors

-> IO JITCompileCallbackManager

Create a new JITCompileCallbackManager.

When the callback manager is no longer needed, it should be freed using disposeJITCompileCallbackManager.

disposeJITCompileCallbackManager :: JITCompileCallbackManager -> IO ()

Dispose of a JITCompileCallbackManager.

withJITCompileCallbackManager

Arguments

:: ShortByteString
target triple

-> Maybe (IO ())
called on compilation errors

-> (JITCompileCallbackManager -> IO a)

-> IO a

Execute a computation using a new JITCompileCallbackManager.

IndirectStubsManagerBuilder

data IndirectStubsManagerBuilder

This is used by CompileOnDemandLayer to manage the stubs created for function definitions that have not yet been compiled.

newIndirectStubsManagerBuilder

Arguments

:: ShortByteString
target triple

-> IO IndirectStubsManagerBuilder

Create a new IndirectStubsManagerBuilder.

When the stubs manager is no longer needed, it should be freed using disposeIndirectStubsManagerBuilder.

disposeIndirectStubsManagerBuilder :: IndirectStubsManagerBuilder -> IO ()

Dispose of an IndirectStubsManagerBuilder.

withIndirectStubsManagerBuilder

Arguments

:: ShortByteString
target triple

-> (IndirectStubsManagerBuilder -> IO a)

-> IO a

bracket-style wrapper around newIndirectStubsManagerBuilder and disposeIndirectStubsManagerBuilder.