LLVM.Transforms

Description

This module provides an enumeration of the various transformation (e.g. optimization) passes provided by LLVM. They can be used to create a PassManager to, in turn, run the passes on Modules. If you don't know what passes you want, consider instead using CuratedPassSetSpec.

Documentation

data Pass

http://llvm.org/docs/Passes.html#transform-passes A few passes can make use of information in a TargetMachine if one is provided to createPassManager. http://llvm.org/doxygen/classllvm_1_1Pass.html

Constructors

AggressiveDeadCodeElimination

BreakCriticalEdges

CodeGenPrepare
can use a TargetMachine

ConstantPropagation

CorrelatedValuePropagation

DeadCodeElimination

DeadInstructionElimination

DeadStoreElimination

DemoteRegisterToMemory

EarlyCommonSubexpressionElimination

GlobalValueNumbering

Fields

noLoads :: Bool

InductionVariableSimplify

InstructionCombining

JumpThreading

LoopClosedSingleStaticAssignment

LoopInvariantCodeMotion

LoopDeletion

LoopIdiom

LoopInstructionSimplify

LoopRotate

LoopStrengthReduce

LoopUnroll

Fields

loopUnrollThreshold :: Maybe Word

count :: Maybe Word

allowPartial :: Maybe Bool

LoopUnswitch

Fields

optimizeForSize :: Bool

LowerAtomic

LowerInvoke

LowerSwitch

LowerExpectIntrinsic

MemcpyOptimization

PromoteMemoryToRegister

Reassociate

ScalarReplacementOfAggregates

Fields

requiresDominatorTree :: Bool

OldScalarReplacementOfAggregates

Fields

oldScalarReplacementOfAggregatesThreshold :: Maybe Word

useDominatorTree :: Bool

structMemberThreshold :: Maybe Word

arrayElementThreshold :: Maybe Word

scalarLoadThreshold :: Maybe Word

SparseConditionalConstantPropagation

SimplifyLibCalls

SimplifyControlFlowGraph

Sinking

TailCallElimination

AlwaysInline

Fields

insertLifetime :: Bool

ArgumentPromotion

ConstantMerge

FunctionAttributes

FunctionInlining

Fields

functionInliningThreshold :: Word

GlobalDeadCodeElimination

InternalizeFunctions

Fields

exportList :: [String]

InterproceduralConstantPropagation

InterproceduralSparseConditionalConstantPropagation

MergeFunctions

PartialInlining

PruneExceptionHandling

StripDeadDebugInfo

StripDebugDeclare

StripNonDebugSymbols

StripSymbols

Fields

onlyDebugInfo :: Bool

LoopVectorize

Fields

noUnrolling :: Bool

alwaysVectorize :: Bool

SuperwordLevelParallelismVectorize

GCOVProfiler

Fields

emitNotes :: Bool

emitData :: Bool

version :: GCOVVersion

useCfgChecksum :: Bool

noRedZone :: Bool

functionNamesInData :: Bool

AddressSanitizer

AddressSanitizerModule

MemorySanitizer

Fields

trackOrigins :: Bool

ThreadSanitizer

BoundsChecking

defaultLoopVectorize :: Pass

Defaults for the LoopVectorize pass

newtype GCOVVersion

See http://gcc.gnu.org/viewcvs/gcc/trunk/gcc/gcov-io.h?view=markup.

Constructors

GCOVVersion ShortByteString

defaultGCOVProfiler :: Pass

Defaults for GCOVProfiler.

defaultAddressSanitizer :: Pass

Defaults for AddressSanitizer.

defaultAddressSanitizerModule :: Pass

Defaults for AddressSanitizerModule.

defaultMemorySanitizer :: Pass

Defaults for MemorySanitizer.

defaultThreadSanitizer :: Pass

Defaults for ThreadSanitizer.